A separate hierarchy for SimConnectMessages is unnecessary
