+++ CsSimConnect/MessageStream.cs
@@ -15,20 +15,17 @@
  */
 
 using System;
-using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace CsSimConnect
 {
-    public class SimConnectMessageStream<T> : SimConnectObserver<T>, IMessageStream<T>
-        where T : SimConnectMessage
+    public class MessageStream<T> : MessageObserver<T>, IMessageStream<T>
+        where T : class
     {
 
-        private static Logger log = Logger.GetLogger(typeof(SimConnectMessageStream<T>));
+        private static Logger log = Logger.GetLogger(typeof(MessageStream<T>));
 
         public uint MaxSize { get; set; }
 
@@ -36,14 +33,9 @@ namespace CsSimConnect
         private T current = null;
         private bool disposedValue;
 
-        public SimConnectMessageStream(uint sendID, uint queueSize) : base(sendID, true)
+        public MessageStream(uint queueSize) : base(true)
         {
             MaxSize = queueSize;
-        }
-
-        override public void OnNext(SimConnectMessage msg)
-        {
-            base.OnNext(msg);
         }
 
         override public void OnNext(T msg)
@@ -83,9 +75,5 @@ namespace CsSimConnect
             throw new NotImplementedException();
         }
 
-        override public IEnumerator<T> GetEnumerator()
-        {
-            throw new NotImplementedException();
-        }
     }
 }
